Add workaround for https://github.com/pyinvoke/invoke/issues/459
Currently fabric2 has a problem with `cd` context manager and `sudo` command. To work around the issue
I implemented a customcontext manager that keeps the cd paths as an array (stack). If stack contains items
then the top path is fed into fabric.cd when running commands with use_sudo=False. If use_sudo is True then
we modify the command executed and wrap it into `bash -c "cd {path} && {command}"`. This should be reverted
to rely on builtin `cd` after the issue gets resolved in upstream.

Also:

- Implement hide as context manager on fabric2